use LiveData subscription pattern for the selection steps
