+++ app/src/main/java/org/sagebionetworks/research/mpower/inject/MPowerApplicationModule.java
@@ -6,23 +6,17 @@ import android.content.Context;
 import org.sagebionetworks.research.mpower.authentication.ExternalIdSignInActivity;
 import org.sagebionetworks.research.mpower.history.HistoryFragment;
 import org.sagebionetworks.research.mpower.insights.InsightsFragment;
-import org.sagebionetworks.research.mpower.tracking.TrackingFragment;
+import org.sagebionetworks.research.mpower.tracking.TrackingTabFragment;
 import org.sagebionetworks.research.mpower.MainActivity;
 import org.sagebionetworks.research.mpower.profile.ProfileFragment;
 import org.sagebionetworks.research.mpower.tracking.TrackingMenuFragment;
-import org.sagebionetworks.research.mpower.tracking.fragment.AddNoteFragment;
 import org.sagebionetworks.research.mpower.tracking.fragment.DurationFragment;
 import org.sagebionetworks.research.mpower.tracking.fragment.SymptomAddNoteFragment;
 import org.sagebionetworks.research.mpower.tracking.fragment.SymptomLoggingFragment;
 import org.sagebionetworks.research.mpower.tracking.fragment.SymptomSelectionFragment;
 import org.sagebionetworks.research.mpower.tracking.fragment.TimePickerFragment;
 import org.sagebionetworks.research.mpower.tracking.fragment.TriggersLoggingFragment;
-import org.sagebionetworks.research.mpower.tracking.fragment.SelectionFragment;
 import org.sagebionetworks.research.mpower.tracking.fragment.TriggersSelectionFragment;
-import org.sagebionetworks.research.mpower.tracking.view_model.SymptomTrackingTaskViewModel;
-import org.sagebionetworks.research.mpower.tracking.view_model.configs.SimpleTrackingItemConfig;
-import org.sagebionetworks.research.mpower.tracking.view_model.logs.SymptomLog;
-import org.sagebionetworks.research.presentation.recorder.service.RecorderService;
 
 import dagger.Binds;
 import dagger.Module;
@@ -55,7 +49,7 @@ public interface MPowerApplicationModule {
     InsightsFragment contributeInsightsFragmentInjector();
 
     @ContributesAndroidInjector
-    TrackingFragment contributeTrackingFragmentInjector();
+    TrackingTabFragment contributeTrackingFragmentInjector();
 
     @ContributesAndroidInjector
     SymptomAddNoteFragment contributeSymptomAddNoteFragmentInjector();
